requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation & Settings
pydantic>=2.7.0
pydantic-settings>=2.1.0

# Image Processing (только чтение изображений, без CV-пайплайна)
opencv-python-headless>=4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# MRZ parsing (оставляем для fallback парсинга, если VLM не справится)
mrz>=0.4.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache & Queue
redis==5.0.1
celery==5.3.6

# Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
httpx>=0.27.0

# Monitoring
prometheus-client==0.19.0
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# VLM - Google Gemini API
google-generativeai>=0.8.0
